+++ src/messages/EmptyMessage.ts
@@ -1,0 +1,8 @@
+import { Builder } from "../boc/Builder";
+import { Message } from "./Message";
+
+export class EmptyMessage implements Message {
+    writeTo(builder: Builder) {
+        // Nothing to do
+    }
+}