+++ src/messages/Message.ts
@@ -1,0 +1,5 @@
+import { Writable } from "../boc/Writable";
+
+export interface Message extends Writable {
+
+}